Cleanup duplicate API methods
