+++ appinfo/routes.php
@@ -24,11 +24,6 @@
 return [
 	'routes' => [
 		['name' => 'page#index', 'url' => '/', 'verb' => 'GET'],
-		['name' => 'page#get', 'url' => '/announcement', 'verb' => 'GET'],
-		['name' => 'page#add', 'url' => '/announcement', 'verb' => 'POST'],
-		['name' => 'page#delete', 'url' => '/announcement/{id}', 'verb' => 'DELETE'],
-		['name' => 'page#removeNotifications', 'url' => '/announcement/{id}/notifications', 'verb' => 'DELETE'],
-		['name' => 'page#searchGroups', 'url' => '/groups', 'verb' => 'GET'],
 	],
 	'ocs' => [
 		['name' => 'API#get', 'url' => '/api/v1/announcements', 'verb' => 'GET'],
